Good work on menu system
Removed the animations for the options menu because its dumb, instead, it will just appear and disappear on button click. Also, made it so the player cannot move / rotate while in a submenu (ie: the options). Good stuff, could not figure out how to disable player gravity so they stay suspended midair when in the menu though.